Move time formatting utils to shared
